HW3 - Fixed

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserMealRepositoryImpl.java
@@ -33,14 +33,12 @@ public class JdbcUserMealRepositoryImpl implements UserMealRepository {
 
     private SimpleJdbcInsert insertMeal;
 
-    private SimpleJdbcInsert insertUsersMeals;
 
     @Autowired
     public JdbcUserMealRepositoryImpl(DataSource dataSource) {
         this.insertMeal = new SimpleJdbcInsert(dataSource)
                 .withTableName("MEALS")
                 .usingGeneratedKeyColumns("id");
-        this.insertUsersMeals = new SimpleJdbcInsert(dataSource).withTableName("USERS_MEALS");
     }
 
     @Override
@@ -49,49 +47,36 @@ public class JdbcUserMealRepositoryImpl implements UserMealRepository {
                 .addValue("id", userMeal.getId())
                 .addValue("dateTime", userMeal.getDateTime())
                 .addValue("description", userMeal.getDescription())
-                .addValue("calories", userMeal.getCalories());
+                .addValue("calories", userMeal.getCalories())
+                .addValue("userId", userId);
 
         if(userMeal.isNew()){
             Number newKey = insertMeal.executeAndReturnKey(map);
             userMeal.setId(newKey.intValue());
 
-            MapSqlParameterSource idMap = new MapSqlParameterSource()
-                    .addValue("mealId", userMeal.getId())
-                    .addValue("userId", userId);
-            insertUsersMeals.execute(idMap);
         } else {
-            if(isMealOwnedByUser(userMeal.getId(), userId)) {
-                namedParameterJdbcTemplate.update(
-                        "UPDATE meals SET date_time=:dateTime, description=:description, calories=:calories WHERE id=:id", map);
-            } else {
-                return null;
-            }
+                if(namedParameterJdbcTemplate.update(
+                        "UPDATE meals SET date_time=:dateTime, description=:description, calories=:calories WHERE id=:id AND user_id=:userId", map) == 0){
+                    return null;
+                }
         }
         return userMeal;
     }
 
     @Override
     public boolean delete(int id, int userId) {
-        if(isMealOwnedByUser(id, userId)) {
-            return jdbcTemplate.update("DELETE FROM meals WHERE id=?", id) != 0;
-        }else {
-            return false;
-        }
+        return jdbcTemplate.update("DELETE FROM meals WHERE id=? AND user_id =?", id, userId) != 0;
     }
 
     @Override
     public UserMeal get(int id, int userId) {
-        if (isMealOwnedByUser(id, userId)) {
-            List<UserMeal> userMeals = jdbcTemplate.query("SELECT * FROM meals WHERE id=?", ROW_MAPPER, id);
+        List<UserMeal> userMeals = jdbcTemplate.query("SELECT * FROM meals WHERE id=? AND user_id=?", ROW_MAPPER, id, userId);
             return DataAccessUtils.singleResult(userMeals);
-        } else {
-            return null;
-        }
     }
 
     @Override
     public List<UserMeal> getAll(int userId) {
-        return jdbcTemplate.query("SELECT * FROM meals WHERE id IN (SELECT meal_id FROM users_meals WHERE user_id = ?) ORDER BY date_time DESC", ROW_MAPPER, userId);
+        return jdbcTemplate.query("SELECT * FROM meals WHERE user_id = ? ORDER BY date_time DESC", ROW_MAPPER, userId);
     }
 
     @Override
@@ -99,8 +84,4 @@ public class JdbcUserMealRepositoryImpl implements UserMealRepository {
         return null;
     }
 
-    private boolean isMealOwnedByUser (int id, int userId){
-        return userId == jdbcTemplate.queryForObject("SELECT user_id FROM users_meals WHERE meal_id=?",
-               Integer.class, id) ? true : false;
-    }
 }
